+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 3.2.1 - 2017-06-24
+
+### Changes
+
+Added Extensions to CallContractOperation, CreateContractOperation, SidechainBurnOperation, SidechainIssueOperation
+
+### Tests
+Updated tests according to changed methods
+
 ## 3.2 - 2017-06-11
 
 ### Changes